🛠️ refactor: 팔로우 페이지 옵션버튼, 리스트 분리 (인증 필요한 api라서 useSuspenseQuery 사용 시 서버에서 한번 호출될 때 오류 발생하여 일단 useQuery 유지) (#153)
